Maps are supported in cells (!)
The syntax is a little weird: `field:key{delimiter}value` , where field
is the name of the field to use, key is the type of keys, value is the
type of values, and delimiter is the string that separates keys from
values and also separates entries (often just a comma). This currently
uses a LinkedHashMap implementation, using code from SquidLib's Maker
class. The TabLabTools class is generated and can be swapped for one
that uses another Map type.