Some cleanup; makes JAR work again
In the last commit, when the project was packaged into a standalone
(shaded) JAR, the TabLabTools.txt resource wasn't packaged correctly or
was otherwise unavailable. Now it is available in a JAR and in the
tests, with a different way of loading it. The batch file that runs this
on Windows is also updated to better handle files with spaces in their
names (I think).