Minor change so empty arrays can be written
Before, if a cell was empty and its type was some kind of array, it
would contain one element of the default value for that array type. As
an example, it would make an array of only "" for a String array cell
with empty contents. Now, an empty array of the correct type is produced
instead, which seems more like the intended result.