Format change to let files be normal TSV
The package info made it kind of a hassle to copy generated code into a
project, so it wasn't really a good thing to have. Now the package is
always "generated" and it's assumed your IDE can rename the package when
you copy it in. This lets the input files be standard TSV, which Github
can display nicely, hooray!